feat: init repo with readme, license, and flake
MPL 2.0 was picked to provide the most flexibility for people who would
simply like to use the code, while maintaining attribution. The wrapping
clause was written to prevent commercialization, as this tool in theory
could do all domain-specific work for someone looking to get rich easy.

The README is written to remind myself as much as others what this
project is meant to be and how to make it. As it grows, it should be
revised for consumer purposes.

The flake is a minimum boilerplate to create a NixOS module with flag to
enable.